# [PERF] speedup stock forecast and picking list view