--- a/addons/stock/report/stock_forecasted.py
+++ b/addons/stock/report/stock_forecasted.py
@@ -51,8 +51,8 @@ class StockForecasted(models.AbstractModel):
 
     def _move_confirmed_domain(self, product_template_ids, product_ids, wh_location_ids):
         in_domain, out_domain = self._move_domain(product_template_ids, product_ids, wh_location_ids)
-        out_domain += [('state', 'not in', ['draft', 'cancel', 'done'])]
-        in_domain += [('state', 'not in', ['draft', 'cancel', 'done'])]
+        out_domain += [('state', 'in', ['waiting', 'confirmed', 'partially_available', 'assigned'])]
+        in_domain += [('state', 'in', ['waiting', 'confirmed', 'partially_available', 'assigned'])]
         return in_domain, out_domain
 
     def _get_report_header(self, product_template_ids, product_ids, wh_location_ids):