@robodoo 2025-04-23
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/207022.png)](https://mergebot.odoo.com/odoo/odoo/pull/207022)