@gurneyalex 2025-04-24
This seems to be the same patch as https://github.com/odoo/odoo/pull/109193/files cc @Whenrow ;-)