@Whenrow 2025-04-24
@gurneyalex It seems we rolled-back the previous version in https://github.com/odoo/odoo/commit/0a43740481c098f076c3aefa550092d6b960be88 without noticing it.
robodoo r+